Updated ParameterFile class
Added functions for reading and writing parameter lists to files, and a function for returning parameter list as a disctionary of names and values. Added type checking to ParameterFile.Parameter class that should be able to check any type given no matter if it is a built in or third party object.