Updated newDir function
Updated newDir function to use os.path methods to normalise the path and allow the function to not be OS specific. Also added arguments to determine when the function will create a new folder with a number if the given path already exists.

Also removed extra whitespace from the file.